Rename authorization (old tastypie-style name) to permissions (DRF-style name) (bug 1197825)
